x x64 bits build is now working

+++ Externals/cmdline/cmdline.h
@@ -524,7 +524,7 @@ public:
   void parse_check(const std::vector<std::string> &args){
     if (!options.count("help"))
       add("help", '?', "print this message");
-    check(args.size(), parse(args));
+    check((int) args.size(), parse(args));
   }
 
   void parse_check(int argc, char *argv[]){
